requirements.txt
# Core app framework
streamlit==1.39.0

# Text embeddings
sentence-transformers==2.7.0

# Vector database
faiss-cpu==1.8.0.post1

# PDF processing
pdfplumber==0.11.0
PyPDF2==3.0.1

# Web content ingestion
requests==2.31.0
beautifulsoup4==4.12.3

# Data handling
numpy==1.26.4
pandas==2.2.2
pickle5==0.0.12

# Gemini API client (Google Generative AI)
google-generativeai==0.6.0
python-dotenv==1.0.1

# Langchain utilities (for text splitting)
langchain==0.2.5
langchain-text-splitters==0.2.2

# Misc utilities
tqdm==4.66.4
regex==2024.5.15
